Allow the assignment to compile
Accidentally added /src folder that contains assignment to .gitignore in
a previous commit. This commit deletes the .gitignore file and again
allows the /src folder to be tracked.

The program now compiles, and the SimplePlayer class is largely complete.
The next question is what sort of generic list should be used to store
players, which I should ask in my tutorial tomorrow.