Implement Callback and some of GameEngine
I think GameEngineCallback is pretty much completed, so all that is left
is to finish the GameEngine logic. The rollPlayer method seems fine,
but the rollHouse method needs work, specifically to implement the two
notes given in the implementation.

Perhaps consider implementing unit testing with JUnit as per Karl's
Medium article?