refactor GameEngineImpl to reduce code duplication
I think further refactoring can be done since the rollPlayer and
rollHouse methods are still very similar, but will have to ask someone
about that.

The majority of changes in this commit are refactoring GameEngine. The
next thing to do would be the applyResult method in GameEngine, and then
I can start writing test cases.